servers/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adapters>=1.2.0",
    "google-genai>=1.30.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.6.5",
    "mcp>=1.13.0",
    "python-dotenv>=1.1.1",
    "google-auth-oauthlib>=1.2.2",
]
